+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,9 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.AndroidComponents"
         tools:targetApi="31">
-
+        <activity
+            android:name=".activity.ServicesActivity"
+            android:exported="false" />
         <activity
             android:name=".ComposeActivity"
             android:exported="false"
