+++ app/src/main/res/layout/activity_main.xml
@@ -24,35 +24,9 @@
         tools:visibility="visible" />
 
     <Button
-        android:id="@+id/buttonStartTimerService"
+        android:id="@+id/buttonStartServicesActivity"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="@string/start_timer_service" />
-
-    <Button
-        android:id="@+id/buttonBindTimerService"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/bind_timer_service" />
-
-    <Button
-        android:id="@+id/buttonStartTimer"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/start_timer"
-        android:visibility="gone" />
-
-    <Button
-        android:id="@+id/buttonBindRemoteService"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/bind_remote_service" />
-
-    <Button
-        android:id="@+id/buttonCallRemoteService"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/call_remote_service"
-        android:visibility="gone" />
+        android:text="@string/start_services_activity" />
 
 </LinearLayout>
