+++ app/src/main/res/layout/activity_services.xml
@@ -1,0 +1,41 @@
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/main"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:gravity="center"
+    android:orientation="vertical"
+    android:padding="16dp">
+
+    <Button
+        android:id="@+id/buttonStartTimerService"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/start_timer_service" />
+
+    <Button
+        android:id="@+id/buttonBindTimerService"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/bind_timer_service" />
+
+    <Button
+        android:id="@+id/buttonStartTimer"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/start_timer"
+        android:visibility="gone" />
+
+    <Button
+        android:id="@+id/buttonBindRemoteService"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/bind_remote_service" />
+
+    <Button
+        android:id="@+id/buttonCallRemoteService"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/call_remote_service"
+        android:visibility="gone" />
+
+</LinearLayout>
